Se implementa el metodo del delete

+++ src/main/java/co/com/crud/requirement/persistence/crud/RequirementCrudRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface RequirementCrudRepository extends CrudRepository<RequirementEntity, Long> {
+public interface RequirementCrudRepository extends CrudRepository<RequirementEntity, Integer> {
     Optional<RequirementEntity> findById(Integer id);
 }
